Graph + zlib proper
zlib dependency was in the wrong place. Instead of being a dependency for the
executable, it was properly moved to being a dependency for the object file.

I created a dot file "graph.gv" that creates an illustration of the dependency
graph. The graph can be built with:
dot -Tpng graph.gv -o graph.png